Remove extra pointers
In Go, *[]mytype means "a pointer to a slice of mytype." It's pretty unusual to want that. A slice is a pointer to a backing array, a length, and a capacity, so you only need *[]mytype if for some reason you need the receiving function to modify your slice in place. On the other hand, []*mytype is a pretty common need because it lets you use objects even if the slice they are in changes. Conceptually, Python lists are []*PyObject in Go terms.